+++ Ex1/Program.cs
@@ -11,33 +11,9 @@ namespace Ex1
         static void Main(string[] args)
         {
             var number1 = new Telephone("999", "999999");
-            Console.WriteLine(number1.OutputPhone());
+            Console.WriteLine(number1.GetPhone());
             var number2 = new Telephone(null, "999999");
-            Console.WriteLine(number2.OutputPhone());
-        }
-    }
-    
-    public class Telephone
-    {
-        public readonly string PhoneCode;
-        public readonly string PhoneNubmer;
-
-        public Telephone(string phoneCode, string phoneNumber)
-        {
-            PhoneCode = phoneCode;
-            PhoneNubmer = phoneNumber;
-        }
-
-        public string OutputPhone()
-        {
-            if (PhoneCode == null)
-            {
-                return PhoneNubmer;
-            }
-            else
-            {
-                return '(' + PhoneCode + ')' + PhoneNubmer;
-            }
+            Console.WriteLine(number2.GetPhone());
         }
     }
 }
